# Use uint16 for storing count