A slice of uint16s fit more snugly in L1 cache and results in fewer misses.

Unfortunately, this limits the length of strings that can be used
to 65536.

Benchmark:
```
name              old time/op  new time/op  delta
Simple/ASCII-4     379ns ± 2%   330ns ± 2%  -13.03%  (p=0.000 n=10+10)
Simple/French-4    692ns ± 2%   617ns ± 2%  -10.75%  (p=0.000 n=10+10)
Simple/Nordic-4   1.28µs ± 1%  1.16µs ± 4%   -9.64%  (p=0.000 n=8+10)
Simple/Tibetan-4  1.16µs ± 3%  1.05µs ± 1%   -9.25%  (p=0.000 n=10+10)
```